v0.9.1 minor bug fix opening bam on wrong chromosome. Added option to use GL backend on linux, not just EGL. Fixed build for termux

+++ src/plot_controls.cpp
@@ -1625,7 +1625,7 @@ namespace Manager {
                     redraw = true;
                 } else if (key >= GLFW_KEY_1 && key <= GLFW_KEY_9 && mouseOverTileIndex >= 0) {
                     int num_idx = key - (int)GLFW_KEY_1;
-                    if (currentVarTrack->multiLabels.empty() || currentVarTrack->blockStart + mouseOverTileIndex > currentVarTrack->multiLabels.size()) {
+                    if (currentVarTrack->multiLabels.empty() || currentVarTrack->blockStart + mouseOverTileIndex > (int)currentVarTrack->multiLabels.size()) {
                         return;
                     }
                     Utils::Label &lbl = currentVarTrack->multiLabels[currentVarTrack->blockStart + mouseOverTileIndex];
